Removed unnecessary fuseable classes
