+++ reactor-core/src/main/java/reactor/core/publisher/FluxContextWriteRestoringThreadLocals.java
@@ -22,6 +22,7 @@ import java.util.function.Function;
 import io.micrometer.context.ContextSnapshot;
 import org.reactivestreams.Subscription;
 import reactor.core.CoreSubscriber;
+import reactor.core.Fuseable;
 import reactor.core.Fuseable.ConditionalSubscriber;
 import reactor.util.annotation.Nullable;
 import reactor.util.context.Context;
@@ -42,7 +43,7 @@ final class FluxContextWriteRestoringThreadLocals<T> extends FluxOperator<T, T> 
 		Context c = doOnContext.apply(actual.currentContext());
 
 		try (ContextSnapshot.Scope ignored = ContextPropagation.setThreadLocals(c)) {
-			source.subscribe(new FuseableContextWriteRestoringThreadLocalsSubscriber<>(actual, c));
+			source.subscribe(new ContextWriteRestoringThreadLocalsSubscriber<>(actual, c));
 		}
 	}
 
@@ -53,7 +54,7 @@ final class FluxContextWriteRestoringThreadLocals<T> extends FluxOperator<T, T> 
 		return super.scanUnsafe(key);
 	}
 
-	static final class FuseableContextWriteRestoringThreadLocalsSubscriber<T>
+	static class ContextWriteRestoringThreadLocalsSubscriber<T>
 			implements ConditionalSubscriber<T>, InnerOperator<T, T> {
 
 		final CoreSubscriber<? super T>        actual;
@@ -63,7 +64,7 @@ final class FluxContextWriteRestoringThreadLocals<T> extends FluxOperator<T, T> 
 		Subscription s;
 
 		@SuppressWarnings("unchecked")
-		FuseableContextWriteRestoringThreadLocalsSubscriber(CoreSubscriber<? super T> actual, Context context) {
+		ContextWriteRestoringThreadLocalsSubscriber(CoreSubscriber<? super T> actual, Context context) {
 			this.actual = actual;
 			this.context = context;
 			if (actual instanceof ConditionalSubscriber) {
@@ -170,6 +171,48 @@ final class FluxContextWriteRestoringThreadLocals<T> extends FluxOperator<T, T> 
 					     ContextPropagation.setThreadLocals(context)) {
 				s.cancel();
 			}
+		}
+	}
+
+	static final class FuseableContextWriteRestoringThreadLocalsSubscriber<T>
+			extends ContextWriteRestoringThreadLocalsSubscriber<T>
+			implements Fuseable.QueueSubscription<T> {
+
+		FuseableContextWriteRestoringThreadLocalsSubscriber(
+				CoreSubscriber<? super T> actual, Context context) {
+			super(actual, context);
+		}
+
+		// Required for
+		// FuseableBestPracticesTest.coreFuseableSubscribersShouldNotExtendNonFuseableOnNext
+		@Override
+		public void onNext(T t) {
+			super.onNext(t);
+		}
+
+		@Override
+		public T poll() {
+			throw new UnsupportedOperationException("Nope");
+		}
+
+		@Override
+		public int requestFusion(int requestedMode) {
+			return Fuseable.NONE;
+		}
+
+		@Override
+		public int size() {
+			throw new UnsupportedOperationException("Nope");
+		}
+
+		@Override
+		public boolean isEmpty() {
+			throw new UnsupportedOperationException("Nope");
+		}
+
+		@Override
+		public void clear() {
+			throw new UnsupportedOperationException("Nope");
 		}
 	}
 }
